requirements.txt
pip~=21.1.3
discord.py~=1.7.3,<2.0.0
matplotlib~=3.4.2
numpy~=1.21.0
